+++ substrait_consumer/functional/queries/ibis_expressions/comparison_functions_expr.py
