+++ substrait_consumer/functional/queries/sql/boolean_functions_sql.py
