+++ substrait_consumer/tests/functional/extension_functions/test_boolean_functions.py
@@ -4,16 +4,16 @@ import duckdb
 from ibis.expr.types.relations import Table
 from ibis_substrait.tests.compiler.conftest import *
 
-from tests.functional.approximation_tests import AGGREGATE_FUNCTIONS
-from tests.parametrization import custom_parametrization
-from tests.functional.common import substrait_function_test
+from substrait_consumer.functional.boolean_configs import AGGREGATE_FUNCTIONS, SCALAR_FUNCTIONS
+from substrait_consumer.functional.common import load_custom_duckdb_table, substrait_function_test
+from substrait_consumer.parametrization import custom_parametrization
 
 
 @pytest.mark.usefixtures("prepare_tpch_parquet_data")
-class TestApproximationFunctions:
+class TestBooleanFunctions:
     """
     Test Class verifying different consumers are able to run substrait plans
-    that include substrait approximation functions.
+    that include substrait boolean functions.
     """
 
     @staticmethod
@@ -22,16 +22,22 @@ class TestApproximationFunctions:
         cls = request.cls
 
         cls.db_connection = duckdb.connect()
-        cls.db_connection.execute("INSTALL substrait")
-        cls.db_connection.execute("LOAD substrait")
+        cls.db_connection.execute("install substrait")
+        cls.db_connection.execute("load substrait")
+        load_custom_duckdb_table(cls.db_connection)
+        cls.table_t = ibis.table(
+            [("a", dt.int32), ("b", dt.int32), ("c", dt.boolean), ("d", dt.boolean)],
+            name="t",
+        )
+
         cls.created_tables = set()
 
         yield
 
         cls.db_connection.close()
 
-    @custom_parametrization(AGGREGATE_FUNCTIONS)
-    def test_approximation_functions(
+    @custom_parametrization(SCALAR_FUNCTIONS + AGGREGATE_FUNCTIONS)
+    def test_boolean_functions(
         self,
         test_name: str,
         file_names: Iterable[str],
@@ -39,7 +45,7 @@ class TestApproximationFunctions:
         ibis_expr: Callable[[Table], Table],
         producer,
         consumer,
-        partsupp,
+        partsupp
     ) -> None:
         substrait_function_test(
             self.db_connection,
@@ -49,5 +55,6 @@ class TestApproximationFunctions:
             ibis_expr,
             producer,
             consumer,
-            partsupp
+            partsupp,
+            self.table_t,
         )
